Bugfixes for service table.

+++ src/stack/service.h
@@ -232,11 +232,11 @@ struct sock *service_find_sock(struct service_id *srvid, int protocol);
 
 void service_entry_hold(struct service_entry *se);
 void service_entry_put(struct service_entry *se);
-int service_entry_print(struct service_entry *se, char *buf, int buflen);
+int service_entry_print(struct service_entry *se, char *buf, size_t buflen);
 
 void service_table_read_lock(void);
 void service_table_read_unlock(void);
-int __service_table_print(char *buf, int buflen);
-int service_table_print(char *buf, int buflen);
+int __service_table_print(char *buf, size_t buflen);
+int service_table_print(char *buf, size_t buflen);
 
 #endif /* _SERVICE_H_ */
